+++ src/pathtracer/pathtracer.cpp
@@ -19,7 +19,7 @@ PathTracer::PathTracer() {
 PathTracer::~PathTracer() {
 }
 
-void PathTracer::set_frame_size(uint16_t width, uint16_t height) {
+void PathTracer::set_frame_size(uint32_t width, uint32_t height) {
   sampleBuffer.resize(width, height);
 }
 
@@ -27,8 +27,8 @@ void PathTracer::clear() {
   bvh = NULL;
 }
 
-void PathTracer::write_to_framebuffer(HDRImageBuffer &buffer, ImageBuffer &framebuffer, uint16_t x0,
-                                      uint16_t y0, uint16_t x1, uint16_t y1) {
+void PathTracer::write_to_framebuffer(HDRImageBuffer &buffer, ImageBuffer &framebuffer, uint32_t x0,
+                                      uint32_t y0, uint32_t x1, uint32_t y1) {
   buffer.toColor(framebuffer, x0, y0, x1, y1);
 }
