+++ src/pathtracer/ray.h
@@ -21,8 +21,8 @@ struct Ray {
 
   Vector3D o;  ///< origin
   Vector3D d;  ///< direction
-  double min_t; ///< treat the ray as a segment (ray "begin" at min_t)
-  double max_t; ///< treat the ray as a segment (ray "ends" at max_t)
+  float min_t; ///< treat the ray as a segment (ray "begin" at min_t)
+  float max_t; ///< treat the ray as a segment (ray "ends" at max_t)
 
   Vector3D inv_d;  ///< component wise inverse
 
@@ -45,12 +45,12 @@ struct Ray {
    * \param max_t max t value for the ray (if it's actually a segment)
    * \param depth depth of the ray
    */
-    HOST_DEVICE Ray(const Vector3D o, const Vector3D d, double max_t, int depth = 0);
+    HOST_DEVICE Ray(const Vector3D o, const Vector3D d, float max_t, int depth = 0);
 
   /**
    * Returns the point t * |d| along the ray.
    */
-  inline Vector3D at_time(double t) const { return o + t * d; }
+  inline Vector3D at_time(float t) const { return o + t * d; }
 };
 
 // structure used for logging rays for subsequent visualization
